Move dashboard tests to separate file
